requirements.txt
opencv-python
streamlit
numpy
pillow
